# agent/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meyaar Error Analysis Agent API",
    "description": "Agentic interpretation layer on top of the PostGIS rule engine.",
    "version": "0.4.0"
  },
  "paths": {
    "/api/validation/{run_id}/analyze": {
      "post": {
        "tags": [
          "validation-analysis"
        ],
        "summary": "Trigger Error Analysis for a validation run",
        "operationId": "trigger_analysis_api_validation__run_id__analyze_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/validation/{run_id}/analysis": {
      "get": {
        "tags": [
          "validation-analysis"
        ],
        "summary": "Retrieve Error Analyses for a validation run",
        "operationId": "get_analysis_api_validation__run_id__analysis_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/validation/{run_id}/remediation": {
      "get": {
        "tags": [
          "validation-analysis"
        ],
        "summary": "Retrieve remediation/audit records for a validation run",
        "description": "Remediation decisions + audit for a run. The dashboard can\ndistinguish: action=auto_fix/status=applied (automatically fixed),\naction=human_review/status=pending_review (needs a reviewer),\naction=no_action/status=none (nothing to do), status=failed (rolled back).\nRecords are created by POST /api/validation/{run_id}/analyze.",
        "operationId": "get_remediation_api_validation__run_id__remediation_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/validation/{run_id}/chat": {
      "post": {
        "tags": [
          "validation-analysis"
        ],
        "summary": "Ask a grounded question about a run's engine results",
        "description": "Chat endpoint: answers from the run's stored analyses + summary only.\n\nRequires an LLM key (MEYAAR_LLM_API_KEY) — analysis endpoints work\nwithout one, chat does not.",
        "operationId": "chat_about_run_api_validation__run_id__chat_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/validation/{run_id}/record": {
      "get": {
        "tags": [
          "validation-analysis"
        ],
        "summary": "Retrieve the COMPLETE analysis data for a run (every stored field)",
        "description": "The unfiltered analysis dataset behind the grounded chat.\n\nReturns the whole stored analysis payload(s) (every JSON key the pipeline\npersisted), the complete record of every feature the run touches (all\nlayer attributes + PostGIS measurements), the raw rule-engine findings\nincluding their details text, and the index of every available field name\nwith an example value. Nothing is whitelisted, so a field added to the\npipeline or a layer later is returned with no code change.",
        "operationId": "get_analysis_data_api_validation__run_id__record_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisListResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "summary": {
            "$ref": "#/components/schemas/RunSummary"
          },
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/ErrorAnalysis"
            },
            "type": "array",
            "title": "Analyses"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "summary"
        ],
        "title": "AnalysisListResponse",
        "description": "GET /api/validation/{run_id}/analysis body."
      },
      "AnalyzeResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_errors_analyzed": {
            "type": "integer",
            "title": "Total Errors Analyzed",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "AnalyzeResponse",
        "description": "POST /api/validation/{run_id}/analyze body."
      },
      "ChatRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "question",
          "answer"
        ],
        "title": "ChatResponse"
      },
      "ErrorAnalysis": {
        "properties": {
          "result_id": {
            "type": "integer",
            "title": "Result Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "layer_name": {
            "type": "string",
            "title": "Layer Name"
          },
          "feature_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Id"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "error_type": {
            "type": "string",
            "title": "Error Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "status": {
            "type": "string",
            "pattern": "^(confirmed|candidate|informational|insufficient_context)$",
            "title": "Status"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "cause": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cause"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "human_review_required": {
            "type": "boolean",
            "title": "Human Review Required",
            "default": false
          },
          "related_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Related Features"
          },
          "insufficient_context": {
            "type": "boolean",
            "title": "Insufficient Context",
            "default": false
          },
          "agent_model": {
            "type": "string",
            "title": "Agent Model",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "result_id",
          "run_id",
          "layer_name",
          "rule_id",
          "error_type",
          "severity",
          "status",
          "explanation"
        ],
        "title": "ErrorAnalysis",
        "description": "Agent interpretation of ONE validation result (agent_error_analysis)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RemediationListResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "remediation": {
            "items": {
              "$ref": "#/components/schemas/RemediationRecord"
            },
            "type": "array",
            "title": "Remediation"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RemediationListResponse",
        "description": "GET /api/validation/{run_id}/remediation body."
      },
      "RemediationRecord": {
        "properties": {
          "remediation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "result_id": {
            "type": "integer",
            "title": "Result Id"
          },
          "layer_name": {
            "type": "string",
            "title": "Layer Name"
          },
          "feature_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Id"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "action": {
            "type": "string",
            "pattern": "^(auto_fix|human_review|no_action)$",
            "title": "Action"
          },
          "remediation_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation Type"
          },
          "status": {
            "type": "string",
            "pattern": "^(applied|failed|pending_review|none)$",
            "title": "Status"
          },
          "issue": {
            "type": "string",
            "title": "Issue",
            "default": ""
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "before_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "Before State"
          },
          "after_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "After State"
          },
          "agent_model": {
            "type": "string",
            "title": "Agent Model",
            "default": ""
          },
          "human_review_required": {
            "type": "boolean",
            "title": "Human Review Required",
            "default": false
          },
          "executed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "result_id",
          "layer_name",
          "rule_id",
          "action",
          "status"
        ],
        "title": "RemediationRecord",
        "description": "Audit row for public.agent_remediation_actions.\n\nOne record per (run_id, result_id). Kept JSON-friendly; before/after\nstates are dicts (JSONB in PostgreSQL). The dashboard can distinguish:\naction=auto_fix + status=applied  -> automatically fixed\naction=human_review + status=pending_review -> requires human reviewer\naction=no_action + status=none    -> nothing to do\naction=auto_fix + status=failed   -> execution failed (logged)"
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "total_errors": {
            "type": "integer",
            "title": "Total Errors",
            "default": 0
          },
          "critical_errors": {
            "type": "integer",
            "title": "Critical Errors",
            "default": 0
          },
          "high_errors": {
            "type": "integer",
            "title": "High Errors",
            "default": 0
          },
          "medium_errors": {
            "type": "integer",
            "title": "Medium Errors",
            "default": 0
          },
          "low_errors": {
            "type": "integer",
            "title": "Low Errors",
            "default": 0
          },
          "most_common_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Most Common Error"
          },
          "priority_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Priority Actions"
          },
          "counts_by_rule": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts By Rule"
          },
          "counts_by_layer": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts By Layer"
          },
          "narrative": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunSummary",
        "description": "Run-level rollup returned alongside individual analyses."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}